Hide native dropdown marker on safari
A collapsible admonish on safari gets a native drop-down marker as well as the one provided by mdbook-admonish.

It doesn't look very good - it's a bit out of alignment and out of place.  I checked on chrome and this element isn't even present there.

This adds some CSS to hide that element in safari.